dev: Update

+++ ErogeHelper.AssistiveTouch/MainWindow.xaml.cs
@@ -1,9 +1,9 @@
 ﻿using ErogeHelper.AssistiveTouch.Helper;
-using ErogeHelper.Share;
+using ErogeHelper.AssistiveTouch.NativeMethods;
+using ErogeHelper.IpcChannel;
 using System.Windows;
 using System.Windows.Input;
 using System.Windows.Interop;
-using System.Windows.Threading;
 
 namespace ErogeHelper.AssistiveTouch
 {
@@ -24,7 +24,7 @@ namespace ErogeHelper.AssistiveTouch
             HwndTools.RemovePopupAddChildStyle(Handle);
             User32.SetParent(Handle, App.GameWindowHandle);
             User32.GetClientRect(App.GameWindowHandle, out var rectClient);
-            User32.SetWindowPos(Handle, IntPtr.Zero, 0, 0, rectClient.Width, rectClient.Height, User32.SetWindowPosFlags.SWP_NOZORDER);
+            User32.SetWindowPos(Handle, IntPtr.Zero, 0, 0, (int)rectClient.Width, (int)rectClient.Height, User32.SetWindowPosFlags.SWP_NOZORDER);
 
             var hooker = new GameWindowHooker();
         }
